+++ src/config/config.manager.ts
@@ -1,12 +1,12 @@
 import { Manager } from 'src/utils/manager.utils.js';
 import { flatten, Flatten } from '../utils/flatten.utils.js';
 
-interface Options<C> {
-  getConfig: () => C;
+interface Options<Config> {
+  getConfig: () => Config;
 }
 
-export class ConfigManager<C extends Record<string, unknown>> extends Manager<Flatten<C>> {
-  constructor({ getConfig }: Options<C>) {
+export class ConfigManager<Config extends Record<string, unknown>> extends Manager<Flatten<Config>> {
+  constructor({ getConfig }: Options<Config>) {
     super({ getSource: () => flatten(getConfig()) });
   }
 }