+++ src/utils/manager.utils.ts
@@ -1,19 +1,19 @@
-interface Options<C> {
-  getSource: () => C;
+interface Options<Source> {
+  getSource: () => Source;
 }
 
-export class Manager<C extends Record<string, unknown>> {
-  protected source: C;
+export class Manager<Source extends Record<string, unknown>> {
+  protected source: Source;
 
-  constructor({ getSource }: Options<C>) {
+  constructor({ getSource }: Options<Source>) {
     this.source = getSource();
   }
 
-  public get<K extends keyof C>(key: K): C[K] {
+  public get<Key extends keyof Source>(key: Key) {
     return this.source[key];
   }
 
-  public getOrThrow<K extends keyof C>(key: K): NonNullable<C[K]> {
+  public getOrThrow<Key extends keyof Source>(key: Key) {
     const value = this.source[key];
 
     if (value === null || value === undefined) {